Landing/backend/requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==3.1.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-js-asset==2.0.0
djangorestframework==3.11.1
idna==3.3
oauthlib==3.2.0
Pillow==9.1.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.3.1
urllib3==1.26.9
